pyproject.toml
[project]
name = "agenticdatapipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastembed>=0.7.3",
    "langchain>=1.0.7",
    "langchain-cerebras>=0.7.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-huggingface>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.3",
    "langsmith>=0.4.43",
    "loguru>=0.7.3",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.44",
]
